app/backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
google-cloud-aiplatform==1.48.0
google-cloud-firestore==2.16.0
google-cloud-secret-manager==2.20.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
tenacity==8.2.3
structlog==24.1.0
